requirements.txt
# Core GUI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Markdown Processing
markdown>=3.5.0
pymdown-extensions>=10.5.0
python-markdown-math>=0.8

# Syntax Highlighting
Pygments>=2.17.0

# Export Functionality (PDF export is optional - requires GTK on Windows)
# weasyprint>=60.0  # Uncomment if you want PDF export (requires GTK libraries on Windows)

# File Operations
watchdog>=3.0.0

# Configuration Management
pyyaml>=6.0

# Optional Git Integration
gitpython>=3.1.40
